+++ packages/salesforcedx-utils-vscode/src/commands/index.ts
@@ -11,4 +11,3 @@ import { NotificationService } from './notificationService';
 export const notificationService = NotificationService.getInstance();
 export { NotificationService };
 export { ProgressNotification } from './progressNotification';
-export { TraceFlags } from './traceFlags';